# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Discord Temporary Access Bot API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Grant {
  id: number;
  discordId: string;
  discordUsername: string;
  grantedAt: string;
  expiresAt: string;
  isExpired: boolean;
}

export interface GrantStatus {
  hasActiveGrant: boolean;
  /** @nullable */
  discordId: string | null;
  /** @nullable */
  discordUsername: string | null;
  /** @nullable */
  expiresAt: string | null;
  /** @nullable */
  hoursRemaining: number | null;
}

export type GetDiscordStatusParams = {
discord_id?: string;
};