# lib/api-zod/src/generated/types/grant.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Discord Temporary Access Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface Grant {
  id: number;
  discordId: string;
  discordUsername: string;
  grantedAt: string;
  expiresAt: string;
  isExpired: boolean;
}