# lib/api-zod/src/generated/types/grantStatus.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Discord Temporary Access Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface GrantStatus {
  hasActiveGrant: boolean;
  /** @nullable */
  discordId: string | null;
  /** @nullable */
  discordUsername: string | null;
  /** @nullable */
  expiresAt: string | null;
  /** @nullable */
  hoursRemaining: number | null;
}